requirements.txt
Brave-Core-Service-API-Bindings==0.1
Flask==0.10.1
Flask-SQLAlchemy==2.0
gevent==1.0.1
google-api-python-client==1.3.1
requests==2.4.3
